cases/interface_test/requirement.txt
pytest
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
allure-pytest
pyyaml
requests
jsonpath
rsa
pymysql